requirements.txt
feedparser>=6.0.11
requests>=2.32.0
click>=8.1.7
PyYAML>=6.0.2
tenacity>=8.2.3
python-dateutil>=2.9.0.post0
schedule>=1.2.2
openai>=1.40.0
markdown>=3.5.2
xhtml2pdf>=0.2.11
